Finished unit tests for product, Refactored all entities to use Long instead of long

+++ src/main/resources/db/migration/V1.0.0.00__init.sql
@@ -12,7 +12,7 @@ CREATE TABLE product (
   name        VARCHAR(200)   NOT NULL,
   description VARCHAR(500)   NOT NULL,
   sku         VARCHAR(10)    NOT NULL,
-  price       DOUBLE(20, 2)  NOT NULL,
+  price       DECIMAL(20, 2) NOT NULL,
   PRIMARY KEY (product_id)
 )
   ENGINE = InnoDB
